+++ tests/testthat/test-expr_meta_random_bayes.R
@@ -1,7 +1,7 @@
 # subtitle from meta-analysis -------------------------------------------
 
 testthat::test_that(
-  desc = "expr_meta_bayes works",
+  desc = "expr_meta_random works",
   code = {
     testthat::skip_if(getRversion() < "3.6")
 
@@ -33,18 +33,19 @@ testthat::test_that(
     # subtitle
     set.seed(123)
     results1 <-
-      suppressWarnings(expr_meta_bayes(
+      suppressWarnings(expr_meta_random(
         data = df1,
+        type = "bayes",
         k = 3,
-        messages = TRUE,
         iter = 1000
       ))
 
     # subtitle
     set.seed(123)
     df <-
-      suppressWarnings(expr_meta_bayes(
+      suppressWarnings(expr_meta_random(
         data = df1,
+        type = "bayes",
         iter = 1000,
         output = "dataframe"
       ))